Feat/#55-매매 기능 구현
