+++ ET-core/src/main/java/com/team1/etcore/trade/service/TradeConsumer.java
@@ -1,5 +1,6 @@
 package com.team1.etcore.trade.service;
 
+import com.team1.etcore.trade.dto.Position;
 import com.team1.etcore.trade.dto.QuoteDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,9 +38,29 @@ public class TradeConsumer {
                     .sellAmount(Integer.parseInt(quoteData[32]))
                     .build();
             // 체결 처리
-            tradeService.processTrade(quoteDTO);
+            processTrade(quoteDTO);
         } catch (Exception e) {
             log.error("Kafka 메시지 처리 중 오류 발생: {}", e.getMessage(), e);
+        }
+    }
+
+    public void processTrade(QuoteDTO quoteDTO) {
+        try {
+            // 매수 주문 처리: "orders:BUY:{stockCode}:{tradePrice}"
+            tradeService.processMatching(Position.BUY,
+                    quoteDTO.getStockCode(),
+                    quoteDTO.getBuyPrice(),
+                    quoteDTO.getBuyAmount()
+            );
+            // 매도 주문 처리: "orders:SELL:{stockCode}:{tradePrice}"
+            tradeService.processMatching(Position.SELL,
+                    quoteDTO.getStockCode(),
+                    quoteDTO.getSellPrice(),
+                    quoteDTO.getSellAmount()
+            );
+        } catch (Exception e) {
+            log.error("체결 처리 중 오류 발생: {}", e.getMessage(), e);
+            throw new RuntimeException("체결 처리 실패", e);
         }
     }
 }
