+++ ET-core/src/test/java/com/team1/etcore/trade/service/TradeServiceTest.java
@@ -1,0 +1,10 @@
+package com.team1.etcore.trade.service;
+
+import org.springframework.web.client.RestClient;
+
+import static org.junit.jupiter.api.Assertions.*;
+//
+//class TradeServiceTest {
+//
+//    RestClient restClient = RestClient.create("http")
+//}
